Inject Users store as a singleton
This will be necessary to add unit tests on WebUI project. The interface
is defined in a new project that will expose all the interfaces. That's
why I called it Definitions.

Also UsersRepository was renamed to UsersStore because the old name made
me think of an implementation of the repository pattern. It's not the
case. In addition, I finally chose to name _privateField like that. This
is what I commonly see on GitHub project for C#, notably in .NET Core
files.